% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_a_name.R
\name{draw_a_name}
\alias{draw_a_name}
\title{draw_a_name}
\usage{
draw_a_name(a, b)
}
\arguments{
\item{a}{name your are interested in}

\item{b}{sex you are intersted in}
}
\value{
a graph of the year-by-year popularity of the chosen name for the chosen sex
}
\description{
draw_a_name
}
\examples{
draw_a_name("Vincent", "M")
}
