% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_names.R
\name{draw_names}
\alias{draw_names}
\title{draw_names}
\usage{
draw_names(x)
}
\arguments{
\item{x}{a vector containint the names you wish to get data on}
}
\value{
one graph with the year by year poopularity of the names in the vector parameter
}
\description{
draw_names
}
\examples{
draw_names(c("Diane", "Vincent"))
}
