% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_excel.R
\name{multi_excel}
\alias{multi_excel}
\title{multi_excel}
\usage{
multi_excel(path)
}
\arguments{
\item{path}{name and path of the excel file you wish to read are}
}
\value{
a list of dataframes
}
\description{
multi_excel
}
\examples{
my_csv_reader("/Users/regisbouyoux/Desktop/Data Science X-HEC/R/X_2017/Jour 4")
}
