% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/My_csv_reader.R
\name{my_csv_reader}
\alias{my_csv_reader}
\title{My_csv_reader}
\usage{
my_csv_reader(folder = getwd())
}
\arguments{
\item{folder}{Path of the folder where the files you wish to read is located}
}
\value{
returns a list of dataframes, each list is a file
}
\description{
My_csv_reader
}
\examples{
my_csv_reader("/Users/regisbouyoux/Desktop/Data Science X-HEC/R/X_2017/Jour 4")
}
