% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help.R
\docType{package}
\name{Packagehomework}
\alias{Packagehomework}
\alias{nameofthepackage-package}
\alias{nameofthepackage}
\alias{Packagehomework-package}
\title{What it does}
\description{
This package contains 5 functions : draw_names, draw_a_name, multi_excel, Save_as_csv, My_csv_reader
}
\author{
Regis Bouyoux <regis.bouyoux@polytechnique.edu>
}
