% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Save_as_csv.R
\name{save_as_csv}
\alias{save_as_csv}
\title{Save_as_csv}
\usage{
save_as_csv(dataset, filename)
}
\arguments{
\item{dataset}{a dataset}

\item{filename}{name of the file you want to save your output as}
}
\value{
a dataset saved as a csv file
}
\description{
Save_as_csv
}
\examples{
save_as_csv(iris, "/Users/regisbouyoux/Desktop/Data Science X-HEC/R/X_2017/Jour 4/output.csv")
}
